move code to header to fix build issues
